Add files via upload
About how many hours would you estimate you worked on this problem set? 48-60

Scale of 1 to 5 (5 is Strongly Agree, 1 is Strongly Disagree): This problem set was easy. 3

Scale of 1 to 5 (5 is Strongly Agree, 1 is Strongly Disagree): This problem set was interesting. 3

What is one thing you feel comfortable with in this problem set? Creating the proper output after multiple trials 

What is one thing you feel uncomfortable with in this problem set? pseudocode and trace tables, not sure if I'm writing the steps out properly, (including enough detail)